+++ fvm/fvm_test.go
@@ -5,6 +5,7 @@ import (
 	"encoding/base64"
 	"encoding/hex"
 	"fmt"
+	"github.com/onflow/cadence/runtime"
 	"strconv"
 	"testing"
 
@@ -1554,6 +1555,192 @@ func TestSignatureVerification(t *testing.T) {
 				))
 			})
 		}
+	}
+}
+
+func TestHashing(t *testing.T) {
+
+	t.Parallel()
+
+	rt := fvm.NewInterpreterRuntime()
+
+	chain := flow.Mainnet.Chain()
+
+	vm := fvm.NewVirtualMachine(rt)
+
+	ctx := fvm.NewContext(
+		zerolog.Nop(),
+		fvm.WithChain(chain),
+		fvm.WithCadenceLogging(true),
+	)
+
+	ledger := testutil.RootBootstrappedLedger(vm, ctx)
+
+	hashScript := func(hashName string) []byte {
+		return []byte(fmt.Sprintf(
+			`
+				import Crypto
+				
+				pub fun main(data: [UInt8]): [UInt8] {
+					return Crypto.hash(data, algorithm: HashAlgorithm.%s)
+				}
+			`, hashName))
+	}
+	hashWithTagScript := func(hashName string) []byte {
+		return []byte(fmt.Sprintf(
+			`
+				import Crypto
+				
+				pub fun main(data: [UInt8], tag: String): [UInt8] {
+					return Crypto.hashWithTag(data, tag: tag, algorithm: HashAlgorithm.%s)
+				}
+			`, hashName))
+	}
+
+	data := []byte("some random message")
+	encodedBytes := make([]cadence.Value, len(data))
+	for i := range encodedBytes {
+		encodedBytes[i] = cadence.NewUInt8(data[i])
+	}
+	cadenceData := jsoncdc.MustEncode(cadence.NewArray(encodedBytes))
+
+	// ===== Test Cases =====
+	cases := []struct {
+		Algo    runtime.HashAlgorithm
+		WithTag bool
+		Tag     string
+		Check   func(t *testing.T, result string, scriptErr errors.Error, executionErr error)
+	}{
+		{
+			Algo:    runtime.HashAlgorithmSHA2_256,
+			WithTag: false,
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "68fb87dfba69b956f4ba98b748a75a604f99b38a4f2740290037957f7e830da8", result)
+			},
+		},
+		{
+			Algo:    runtime.HashAlgorithmSHA2_384,
+			WithTag: false,
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "a9b3e62ab9b2a33020e015f245b82e063afd1398211326408bc8fc31c2c15859594b0aee263fbb02f6d8b5065ad49df2", result)
+			},
+		},
+		{
+			Algo:    runtime.HashAlgorithmSHA3_256,
+			WithTag: false,
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "38effea5ab9082a2cb0dc9adfafaf88523e8f3ce74bfbeac85ffc719cc2c4677", result)
+			},
+		},
+		{
+			Algo:    runtime.HashAlgorithmSHA3_384,
+			WithTag: false,
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "f41e8de9af0c1f46fc56d5a776f1bd500530879a85f3b904821810295927e13a54f3e936dddb84669021052eb12966c3", result)
+			},
+		},
+		{
+			Algo:    runtime.HashAlgorithmSHA2_256,
+			WithTag: true,
+			Tag:     "some_tag",
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "cd323b6743f8c72d1b19697aebb0c6c90240c258cd58a2f183c4fde9753143e2", result)
+			},
+		},
+		{
+			Algo:    runtime.HashAlgorithmSHA2_384,
+			WithTag: true,
+			Tag:     "some_tag",
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "bb7c965165ab02e8196cc2ba1159a3212ed94b8ca27b34698a08bd8f1dcfe2cd5806561e3d66fde304f786069fb89094", result)
+			},
+		},
+		{
+			Algo:    runtime.HashAlgorithmSHA3_256,
+			WithTag: true,
+			Tag:     "some_tag",
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "8cd65b26b10c596628357d9ab47c4730228d27db11580ddd584e490107cb9c76", result)
+			},
+		},
+		{
+			Algo:    runtime.HashAlgorithmSHA3_384,
+			WithTag: true,
+			Tag:     "some_tag",
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "8f3d63c56dda4ea9dd638a22bda104b1e272feb11c415ac48f4d3dd1dfd3de5f091a5417501615968d5e3729a74f03fa", result)
+			},
+		},
+		{
+			Algo:    runtime.HashAlgorithmKMAC128_BLS_BLS12_381,
+			WithTag: false,
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "627d7e8fe50384601ca550ceecb61c23e9cbde7feb75ae6b53227f128f2dc3b78b543a044058403e4822f88cb7040d90d588c9e8575f0de3012fe7edaf02b9997a8a5fad234d21b2af359ec3abaeaf4a7ef60e5f04623a983bd5e071f4113678710e910d48ac4d1713073a707ab9057867e0ba32aca6b33010b1d20b8006dd25", result)
+			},
+		},
+		{
+			Algo:    runtime.HashAlgorithmKMAC128_BLS_BLS12_381,
+			WithTag: true,
+			Tag:     "some_tag",
+			Check: func(t *testing.T, result string, scriptErr errors.Error, executionErr error) {
+				require.NoError(t, scriptErr)
+				require.NoError(t, executionErr)
+				require.Equal(t, "dc6889f9ca46803a9c7759068989dfc3cffe632fd991e25f6589603c73b7891e2f4736eebe5248f211bbddaa3d763b1b9318185eaf3ab3bfd6f159f345c3148795e4ff3ad376c98d5616febebcf4520ca2a83dda4be2f98b1ead9fb5a622355305b156e06db173a9e1d7af973b11acc1e714cd3aa0fb367dfaadc5a957b4742b", result)
+			},
+		},
+	}
+	// ======================
+
+	for i, c := range cases {
+		t.Run(fmt.Sprintf("case %d: %s with tag: %v", i, c.Algo, c.WithTag), func(t *testing.T) {
+			code := hashScript(c.Algo.Name())
+			if c.WithTag {
+				code = hashWithTagScript(c.Algo.Name())
+			}
+
+			script := fvm.Script(code)
+
+			if c.WithTag {
+				script = script.WithArguments(
+					cadenceData,
+					jsoncdc.MustEncode(cadence.String(c.Tag)),
+				)
+			} else {
+				script = script.WithArguments(
+					cadenceData,
+				)
+			}
+
+			err := vm.Run(ctx, script, ledger, programs.NewEmptyPrograms())
+
+			byteResult := make([]byte, 0)
+			if err == nil && script.Err == nil {
+				cadenceArray := script.Value.(cadence.Array)
+				for _, value := range cadenceArray.Values {
+					byteResult = append(byteResult, value.(cadence.UInt8).ToGoValue().(uint8))
+				}
+			}
+
+			c.Check(t, hex.EncodeToString(byteResult), script.Err, err)
+		})
 	}
 }
